+++ changelog.md
@@ -3,12 +3,13 @@
 ## TODO
 - Set bank size; show remaining bank in the status bar.
 - Show total cash in play in the status bar if bank is not set.
-- Give and take cash.
+- Show company values.
 
 ## Unreleased
 - Added: Instead of sum, dividends can be now paid as `prev` to repeat the
 previous dividend payment for the company.
 - Added: Automated tests.
+- Added: Player cash can be adjusted with `give` and `take`.
 
 ## 1.0.0 - 2019-06-17
 - First released version!